// backend/be/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ListingStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  SOLD: 'SOLD',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type ListingStatus = (typeof ListingStatus)[keyof typeof ListingStatus]


export const PricingType = {
  FIXED: 'FIXED',
  BIDDING: 'BIDDING'
} as const

export type PricingType = (typeof PricingType)[keyof typeof PricingType]


export const ConditionType = {
  NEW: 'NEW',
  LIKE_NEW: 'LIKE_NEW',
  USED: 'USED',
  REFURBISHED: 'REFURBISHED'
} as const

export type ConditionType = (typeof ConditionType)[keyof typeof ConditionType]


export const BidStatus = {
  VALID: 'VALID',
  OUTBID: 'OUTBID',
  WON: 'WON',
  CANCELLED: 'CANCELLED'
} as const

export type BidStatus = (typeof BidStatus)[keyof typeof BidStatus]


export const SaleType = {
  FIXED: 'FIXED',
  BIDDING: 'BIDDING'
} as const

export type SaleType = (typeof SaleType)[keyof typeof SaleType]


export const OrderStatus = {
  CREATED: 'CREATED',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  REQUIRES_ACTION: 'REQUIRES_ACTION',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
  CANCELLED: 'CANCELLED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentGateway = {
  STRIPE: 'STRIPE',
  RAZORPAY: 'RAZORPAY',
  PAYPAL: 'PAYPAL',
  OTHER: 'OTHER'
} as const

export type PaymentGateway = (typeof PaymentGateway)[keyof typeof PaymentGateway]